agents/analytics/requirements.txt
fastapi
uvicorn
jinja2
httpx
python-dotenv